FINISH EMAIL NOTIFICATION. Update Readme. Rename dbconfig to pjconfig, add 'sender email config' to pjconfig file. Please read README for more information.
