+++ README.md
@@ -5,8 +5,12 @@ Download model-r100-ii from github and paste to models/insightface/models
 cd autoface
 export PYTHONPATH=./models/insightface/deploy/
 
-save db info in dbconfig.py
-in dbconfig.py:
+save db config in pjconfig.py
+save sender email config in pjconfig.py
+
+in pjconfig.py:
+
+  # MongoDB database config
     DBUSERNAME=db_username
     DBPASSWORD=db_password
     DBHOST=db_host
@@ -14,6 +18,12 @@ in dbconfig.py:
     DBNAME=db_name
     COLNAME=collection_name
 
+  # sender email config
+    SENDER_EMAIL=sender_email (gmail)
+    SENDER_EMAIL_PASSWORD=
+    SMTP_SERVER="smtp.gmail.com"
+
+save receiver email list in receiver-email file
 
 # manage config.json
 {
