+++ pjconfig.py
@@ -8,4 +8,3 @@ COLNAME='PredictFaces'  #collection_name
 SENDER_EMAIL="learnstmp.pythonmail@gmail.com"
 SENDER_EMAIL_PASSWORD="Zx9mw(q@dDkwAzf'ky;\p{+P=L(!S_JhWrSCbD88KVj\"[K&sW(Ju;3SM~$w=s[m8"
 SMTP_SERVER="smtp.gmail.com"
-RECEIVER_EMAIL="fake.mail.2043@gmail.com"
